Update kernel config to support emulator WiFi
Enable process and network namespaces in the kernel config as well as
NAT to support WiFi in the emulator. The WiFi solution relies on
namespaces to separate the two ends of the virtual ethernet and WiFi
networks and NAT to do address translation for IPv4 across that virtual
network. This change also enable the drivers needed for virtual WiFI and
ethernet networks.

(cherry picked from commit 5cd4762ac63244fb78188761a57af5e1e242996c)

Change-Id: If5a3b67804df0a4058affb6b033056124f50645b
Signed-off-by: "Bjoern Johansson" <bjoernj@google.com>