MIPS: ranchu: Fix Goldfish PIC initialization and IRQ routing
Goldfish PIC is connected to the CPU IRQ 2 line and should be acting
as a cascaded interrupt controller. Before this change this was not
entirely the case and a workaround was needed in the kernel driver
to route CPU IRQs 0..7 through Goldfish PIC device which was not the
proper way of functioning nor it was following the virtual platform design.

Additionaly, the Goldfish PIC device initial design for masking and
unmasking interrupts was using IRQ number instead of proper IRQ mask
as input values. This is now sorted in QEMU Goldfish PIC implementation
and allows us to use the IRQ generic chip API in the kernel for the
driver, which significantly improves portability.

Device tree needs an updated layout of the Interrupt controller nodes
and its topological connections thus requires adding information about a
CPU interrupt controller which is a parent to the Goldfish PIC and is
supposed to handle interrupts between 0..7 using separate driver in the
kernel.

The final DT layout of the interrupt controller nodes should look as follows:
    goldfish_pic@1f000000 {
            interrupts = <0x2>;
            interrupt-parent = <0x8000>;
            compatible = "google,goldfish-pic", "generic,goldfish-pic";
            phandle = <0x8001>;
            #interrupt-cells = <0x1>;
            interrupt-controller;
            reg = <0x1f000000 0x0 0x1000>;
    };

    cpuintc {
            compatible = "mti,cpu-interrupt-controller";
            phandle = <0x8000>;
            #interrupt-cells = <0x1>;
            interrupt-controller;
    };

Change-Id: I82a9e52b2cb5cb40bb6fa71db9e464b213752829
Signed-off-by: Miodrag Dinic <miodrag.dinic@imgtec.com>